requirements.txt
# Deep Learning

torch
torchvision
torchaudio

# Data Processing

numpy
pandas

# Image Processing

pillow
opencv-python

# Visualization

matplotlib
seaborn

# Model Utilities

scikit-learn
tqdm

# Web Dashboard

streamlit
streamlit-image-comparison 
reportlab

# API

fastapi
uvicorn
python-multipart

# Experiment Tracking

mlflow
